Separate download requests from the TelegramClient
